+++ RealmSwift/SwiftUI.swift
@@ -1048,6 +1048,8 @@ private class ObservableAsyncOpenStorage: ObservableObject {
     var configuration: Realm.Configuration?
     var partitionValue: AnyBSON?
 
+    var setupHasRun = false
+
     // Tracks User State for App for Multi-User Support
     enum AppState {
         case loggedIn(User)
@@ -1148,7 +1150,6 @@ private class ObservableAsyncOpenStorage: ObservableObject {
 
         if let user = app.currentUser {
             appState = .loggedIn(user)
-            asyncOpenForUser(user)
         } else {
             appState = .loggedOut
             asyncOpenState = .waitingForUser
@@ -1253,7 +1254,12 @@ private class ObservableAsyncOpenStorage: ObservableObject {
 
     /// :nodoc:
     public var wrappedValue: AsyncOpenState {
-        storage.asyncOpenState
+        if !storage.setupHasRun {
+            storage.asyncOpen()
+            storage.setupHasRun = true
+        }
+        return storage.asyncOpenState
+
     }
 
     /**
@@ -1290,6 +1296,21 @@ private class ObservableAsyncOpenStorage: ObservableObject {
                  if empty the user configuration will be used.
      - parameter timeout: The maximum number of milliseconds to allow for a connection to
                  become fully established., if empty or `nil` no connection timeout is set.
+
+     - Note: Adding a configuration on this property wrapper will use the file url defined on the `Realm.Configuration` for the
+             synced realm. Keep in mind this may conflict in case you are opening different realm files with different partitions or
+             using different users if the file url is always `default.realm`.
+     - Note: It is recommend to use the configuration obtained from the user `user.configuration(partitionValue:)`
+             when adding a configuration to this property wrapper, and modify any configuration option from it.
+
+             AsyncOpenView()
+                 .environment(\.realmConfiguration, getConfiguration())
+
+             func getConfiguration() -> Realm.Configuration {
+                 var configuration = user.configuration(partitionValue: "partitionValue")
+                 configuration.encryptionKey = getKey()
+                 return configuration
+             }
      */
     public init(appId: String? = nil,
                 configuration: Realm.Configuration? = nil,
@@ -1304,7 +1325,6 @@ private class ObservableAsyncOpenStorage: ObservableObject {
             let bsonValue = AnyBSON(partitionValue: partitionValue)
             if storage.partitionValue != bsonValue {
                 storage.partitionValue = bsonValue
-                storage.asyncOpen()
             }
         }
 
@@ -1315,7 +1335,6 @@ private class ObservableAsyncOpenStorage: ObservableObject {
                 storage.partitionValue = partitionValue
             }
             storage.configuration = configuration
-            storage.asyncOpen()
         }
     }
 }
@@ -1378,7 +1397,11 @@ private class ObservableAsyncOpenStorage: ObservableObject {
 
     /// :nodoc:
     public var wrappedValue: AsyncOpenState {
-        storage.asyncOpenState
+        if !storage.setupHasRun {
+            storage.asyncOpen()
+            storage.setupHasRun = true
+        }
+        return storage.asyncOpenState
     }
 
     /**
@@ -1397,6 +1420,21 @@ private class ObservableAsyncOpenStorage: ObservableObject {
                  if empty the user configuration will be used.
      - parameter timeout: The maximum number of milliseconds to allow for a connection to
                  become fully established, if empty or `nil` no connection timeout is set.
+
+     - Note: Adding a configuration on this property wrapper will use the file url defined on the `Realm.Configuration` for the
+             synced realm. Keep in mind this may conflict in case you are opening different realm files with different partitions or
+             using different users if the file url is always `default.realm`.
+     - Note: It is recommend to use the configuration obtained from the user `user.configuration(partitionValue:)`
+             when adding a configuration to this property wrapper, and modify any configuration option from it.
+
+             AsyncOpenView()
+                 .environment(\.realmConfiguration, getConfiguration())
+
+             func getConfiguration() -> Realm.Configuration {
+                 var configuration = user.configuration(partitionValue: "partitionValue")
+                 configuration.encryptionKey = getKey()
+                 return configuration
+             }
      */
     public init<Partition>(appId: String? = nil,
                            partitionValue: Partition,
@@ -1429,7 +1467,6 @@ private class ObservableAsyncOpenStorage: ObservableObject {
             let bsonValue = AnyBSON(partitionValue: partitionValue)
             if storage.partitionValue != bsonValue {
                 storage.partitionValue = bsonValue
-                storage.asyncOpen()
             }
         }
 
@@ -1440,7 +1477,6 @@ private class ObservableAsyncOpenStorage: ObservableObject {
                 storage.partitionValue = partitionValue
             }
             storage.configuration = configuration
-            storage.asyncOpen()
         }
     }
 }